requirements.txt
alabaster==0.7.12
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
packaging==19.2
Pygments==2.4.2
pyparsing==2.4.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinx-argparse==0.2.5
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.25.3
